livekit-agent/requirements.txt
livekit-agents[openai,silero]~=1.2
python-dotenv~=1.0
certifi~=2024.12
